+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History
 
+## 1.2.0 (2022-12-05)
+
+### Features Added
+- Added OS Patching Parameters to Compute Instance.
+
+### Bugs Fixed
+- Fixed Sweep node not excluding optional input with value `None` in REST object. 
+- Fixed bool test for output in download operation.
+
 ## 1.1.2 (2022-11-21)
 
 ### Features Added